+++ apps/test-suite/tests/SQLite.ts
@@ -1,4 +1,3 @@
-import assert from 'assert';
 import { Asset } from 'expo-asset';
 import * as FS from 'expo-file-system';
 import { Platform } from 'expo-modules-core';
@@ -628,7 +627,6 @@ export function test(t) {
         await db.transactionAsync(async (tx) => {
           await tx.executeSqlAsync('INSERT INTO Users (name) VALUES (?)', [userName]);
           const result = await tx.executeSqlAsync('SELECT * FROM Users LIMIT 1');
-          assert(!isResultSetError(result));
           const currentUser = result.rows[0].name;
           t.expect(currentUser).toEqual('Tim Duncan');
         });
@@ -654,7 +652,6 @@ export function test(t) {
           ]);
 
           const result = await tx.executeSqlAsync('SELECT COUNT(*) FROM Users');
-          assert(!isResultSetError(result));
           const recordCount = result.rows[0]['COUNT(*)'];
           t.expect(recordCount).toEqual(3);
         });
@@ -667,10 +664,16 @@ export function test(t) {
 
           // create table in readOnly transaction
           await db.transactionAsync(async (tx) => {
-            const result = await tx.executeSqlAsync('DROP TABLE IF EXISTS Users;', []);
-            assert(isResultSetError(result));
-            t.expect(result.error).toBeDefined();
-            t.expect(result.error.message).toContain('could not prepare ');
+            let error: Error | null = null;
+            try {
+              await tx.executeSqlAsync('DROP TABLE IF EXISTS Users;', []);
+            } catch (e: unknown) {
+              if (e instanceof Error) {
+                error = e;
+              }
+            }
+            t.expect(error).toBeDefined();
+            t.expect(error.message).toContain('could not prepare ');
           }, true);
         }
       );
@@ -691,7 +694,6 @@ export function test(t) {
         });
         await db.transactionAsync(async (tx) => {
           const result = await tx.executeSqlAsync('SELECT COUNT(*) FROM Users');
-          assert(!isResultSetError(result));
           const recordCount = result.rows[0]['COUNT(*)'];
           t.expect(recordCount).toEqual(1);
         }, true);
@@ -707,7 +709,6 @@ export function test(t) {
 
         await db.transactionAsync(async (tx) => {
           const result = await tx.executeSqlAsync('SELECT COUNT(*) FROM Users');
-          assert(!isResultSetError(result));
           const recordCount = result.rows[0]['COUNT(*)'];
           t.expect(recordCount).toEqual(1);
         }, true);
@@ -723,7 +724,6 @@ export function test(t) {
           );
           // a result-returning pragma
           let result = await tx.executeSqlAsync('PRAGMA table_info(SomeTable);', []);
-          assert(!isResultSetError(result));
           t.expect(result.rows.length).toEqual(2);
           t.expect(result.rows[0].name).toEqual('id');
           t.expect(result.rows[1].name).toEqual('name');
@@ -732,17 +732,10 @@ export function test(t) {
           // a setter/getter pragma
           await tx.executeSqlAsync('PRAGMA user_version = 123;', []);
           result = await tx.executeSqlAsync('PRAGMA user_version;', []);
-          assert(!isResultSetError(result));
           t.expect(result.rows.length).toEqual(1);
           t.expect(result.rows[0].user_version).toEqual(123);
         });
       });
     }); // t.describe('SQLiteAsync')
   }
-}
-
-function isResultSetError(
-  result: SQLite.ResultSet | SQLite.ResultSetError
-): result is SQLite.ResultSetError {
-  return 'error' in result;
 }